Improvements in the Web Services Layers.
Fixing some bad parameteres in the Uploading Images methods.
Improving the via for building a JSON Entity object.
Removing bad checking validations.

+++ ANDROID_Application/OMRGrader_2.0/src/co/edu/udea/android/omrgrader2_0/webservice/IExamImageWebService.java
@@ -12,7 +12,7 @@ import co.edu.udea.android.omrgrader2_0.webservice.model.GraderSession;
  */
 public interface IExamImageWebService {
 
-	public int buildStorageDirectoryPathName(GraderSession graderSession);
+	public long buildStorageDirectoryPathName(GraderSession graderSession);
 
 	public boolean uploadReferenceExamImageFile(
 			Bitmap referenceExamImageBitmap, GraderSession graderSession)
